fixed a test and moved unit tests to data stream module, because these relied on _timestamp field mapper.
