+++ server/src/test/java/org/elasticsearch/cluster/metadata/MetadataCreateDataStreamServiceTests.java
@@ -153,34 +153,6 @@ public class MetadataCreateDataStreamServiceTests extends ESTestCase {
         return MetadataCreateDataStreamService.createDataStream(metadataCreateIndexService, cs, req);
     }
 
-    public void testValidateTimestampFieldMapping() throws Exception {
-        String mapping = generateMapping("@timestamp", "date");
-        validateTimestampFieldMapping("@timestamp", createMapperService(mapping));
-        mapping = generateMapping("@timestamp", "date_nanos");
-        validateTimestampFieldMapping("@timestamp", createMapperService(mapping));
-    }
-
-    public void testValidateTimestampFieldMappingNoFieldMapping() {
-        Exception e = expectThrows(IllegalArgumentException.class,
-            () -> validateTimestampFieldMapping("@timestamp", createMapperService("{}")));
-        assertThat(e.getMessage(),
-            equalTo("[_timestamp] meta field doesn't point to data stream timestamp field [@timestamp]"));
-
-        String mapping = generateMapping("@timestamp2", "date");
-        e = expectThrows(IllegalArgumentException.class,
-            () -> validateTimestampFieldMapping("@timestamp", createMapperService(mapping)));
-        assertThat(e.getMessage(),
-            equalTo("[_timestamp] meta field doesn't point to data stream timestamp field [@timestamp]"));
-    }
-
-    public void testValidateTimestampFieldMappingInvalidFieldType() {
-        String mapping = generateMapping("@timestamp", "keyword");
-        Exception e = expectThrows(IllegalArgumentException.class,
-            () -> validateTimestampFieldMapping("@timestamp", createMapperService(mapping)));
-        assertThat(e.getMessage(), equalTo("the configured timestamp field [@timestamp] is of type [keyword], " +
-            "but [date,date_nanos] is expected"));
-    }
-
     private static MetadataCreateIndexService getMetadataCreateIndexService() throws Exception {
         MetadataCreateIndexService s = mock(MetadataCreateIndexService.class);
         when(s.applyCreateIndexRequest(any(ClusterState.class), any(CreateIndexClusterStateUpdateRequest.class), anyBoolean()))
@@ -204,18 +176,4 @@ public class MetadataCreateDataStreamServiceTests extends ESTestCase {
         return s;
     }
 
-    MapperService createMapperService(String mapping) throws IOException {
-        String indexName = "test";
-        IndexMetadata indexMetadata = IndexMetadata.builder(indexName)
-            .settings(Settings.builder()
-                .put(IndexMetadata.SETTING_VERSION_CREATED, Version.CURRENT)
-                .put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, 1)
-                .put(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 1))
-            .putMapping(mapping)
-            .build();
-        MapperService mapperService =
-            MapperTestUtils.newMapperService(xContentRegistry(), createTempDir(), Settings.EMPTY, indexName);
-        mapperService.merge(indexMetadata, MapperService.MergeReason.MAPPING_UPDATE);
-        return mapperService;
-    }
 }
